Move client to test only target, remove deps
